Create simple_webpage_using _javascript
Prompt the user for their name. Then, here use a pop-up box to welcome the user by name.After that, display the current date on the page in this format: April 30, 2014.  Do not display the time. When i load this page, i will get a current date. And also,  display the last modified date of the document